Reorganized CrawlCtrl unit tests... because I can :p
