+++ test/CrawlCtrl.UnitTests/LineTypes/UnknownDirective/UnknownDirectiveCommentTests.cs
@@ -1,22 +1,23 @@
 using CrawlCtrl.UnitTests.TestData;
 using Xunit;
 
-namespace CrawlCtrl.UnitTests.General;
+namespace CrawlCtrl.UnitTests.LineTypes.UnknownDirective;
 
-public sealed class InvalidLineCommentTests
+public sealed class UnknownDirectiveCommentTests
 {
     [Theory]
     [MemberData(nameof(CommentTestData.OriginalCommentTestData), MemberType = typeof(CommentTestData))]
     public void WHEN_Created_with_comment_THEN_Original_comment_is_set_to_provided_value(string comment, string expectedOriginalComment)
     {
         // Arrange
+        const string directive = "directive";
         const string value = "Some value";
 
         // Act
-        var invalidLine = new InvalidLine(value: value, comment: comment);
+        var unknownDirective = new CrawlCtrl.UnknownDirective(directive: directive, value: value, comment: comment);
 
         // Assert
-        Assert.Equal(expectedOriginalComment, invalidLine.OriginalComment);
+        Assert.Equal(expectedOriginalComment, unknownDirective.OriginalComment);
     }
     
     [Theory]
@@ -24,70 +25,75 @@ public sealed class InvalidLineCommentTests
     public void WHEN_Created_with_comment_THEN_Comment_is_set_to_whitespace_trimmed_value(string comment, string expectedComment)
     {
         // Arrange
+        const string directive = "directive";
         const string value = "Some value";
 
         // Act
-        var invalidLine = new InvalidLine(value: value, comment: comment);
+        var unknownDirective = new CrawlCtrl.UnknownDirective(directive: directive, value: value, comment: comment);
 
         // Assert
-        Assert.Equal(expectedComment, invalidLine.Comment);
+        Assert.Equal(expectedComment, unknownDirective.Comment);
     }
     
     [Fact]
     public void WHEN_Created_with_null_comment_THEN_Original_comment_is_null()
     {
         // Arrange
+        const string directive = "directive";
         const string value = "Some value";
         const string? comment = null;
 
         const string? expectedComment = null;
 
         // Act
-        var invalidLine = new InvalidLine(value: value, comment: comment);
+        var unknownDirective = new CrawlCtrl.UnknownDirective(directive: directive, value: value, comment: comment);
 
         // Assert
-        Assert.Equal(expectedComment, invalidLine.OriginalComment);
+        Assert.Equal(expectedComment, unknownDirective.OriginalComment);
     }
 
     [Fact]
     public void WHEN_Created_with_null_comment_THEN_Comment_is_null()
     {
         // Arrange
+        const string directive = "directive";
         const string value = "Some value";
         const string? comment = null;
 
         const string? expectedComment = null;
 
         // Act
-        var invalidLine = new InvalidLine(value: value, comment: comment);
+        var unknownDirective = new CrawlCtrl.UnknownDirective(directive: directive, value: value, comment: comment);
 
         // Assert
-        Assert.Equal(expectedComment, invalidLine.Comment);
+        Assert.Equal(expectedComment, unknownDirective.Comment);
     }
     
     [Fact]
     public void WHEN_Created_without_comment_THEN_Original_comment_is_null()
     {
         // Arrange
+        const string directive = "directive";
         const string value = "Some value";
         
         // Act
-        var invalidLine = new InvalidLine(value: value, comment: null);
+        var unknownDirective = new CrawlCtrl.UnknownDirective(directive: directive, value: value);
         
         // Assert
-        Assert.Null(invalidLine.OriginalComment);
+        Assert.Null(unknownDirective.OriginalComment);
     }
     
     [Fact]
     public void WHEN_Created_without_comment_THEN_Comment_is_null()
     {
         // Arrange
+        const string directive = "directive";
         const string value = "Some value";
         
         // Act
-        var invalidLine = new InvalidLine(value: value, comment: null);
+        var unknownDirective = new CrawlCtrl.UnknownDirective(directive: directive, value: value);
         
         // Assert
-        Assert.Null(invalidLine.Comment);
+        Assert.Null(unknownDirective.Comment);
     }
 }
